requirements.txt
requests>=2.31.0
google-cloud-bigquery>=3.13.0
anthropic>=0.18.0
gspread>=6.0.0
google-auth>=2.22.0
